Safari-609.1.11.1/LayoutTests/fast/dom/getElementsByTagName-return-type.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that getElementsByTagName*() returns an HTMLCollection.");

shouldBe('document.getElementsByTagName("test").__proto__', 'HTMLCollection.prototype');
shouldBe('document.getElementsByTagNameNS("*", "test").__proto__', 'HTMLCollection.prototype');
shouldBe('document.body.getElementsByTagName("test").__proto__', 'HTMLCollection.prototype');
shouldBe('document.body.getElementsByTagNameNS("*", "test").__proto__', 'HTMLCollection.prototype');
</script>
<script src="../../resources/js-post.js"></script>
<body>
</html>
